Combined space-store and testimonial-store into use-space
useSpace is now a single uniform hook for accessing and updating space metadata and testimonials.

Prior, useSpace depended on the result of useSteps, which itself depended on the result of useTestimonial. This dependency chain was unnessecary and unintuitive.

Also the way fetching and updating work under the hood has changed.

The space and testimonial properties of useSpace now use a default/fallback value if no matching record is in the local database. Only during initialization of useLiveQuery, useSpace returns null for these properties.

updateSpace and updateTestimonial do now upsert the data. So no need to check if the record actually exists.

This seems to be a much cleaner solution, than to seed the database manually, when mounting for the first time and no space, testimonial, or answers exist.